Add files via upload
I added more than just the files that work, I also added the process files. The original code would screenshot the whole screen, use pytesseract to detect the "continue with interruptions" text, and then click at certain coordinates if it found it. It was very clunky and also took about a second for everything to register, rendering the blocker useless. I then tried to use pyautogui's ability to detect if an image is on screen, hence the "interruptionsimg", but I couldn't get that to work. While researching how I could get it to work, I learnt it would also take about a second for it to register. While scrolling through pyautogui's documentation, I found that you could detect pixel colors. I did that, and made it click at the continue with interruptions text when it found a blue pixel (meaning that the link to continue was there) in the continue with interruptions area. The pixel coordinates are hardcoded, so if you want to change the code feel free to use the mousecoordscheck.py file to check where the continue text shows up for you and change the coordinates (you can check where the text is by getting the songsterr popup to show and then running the code and clicking on the continue text).